[#77]/refactor FolderPath 컴포넌트 추출
